% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_manipulation.R
\name{add_list_items}
\alias{add_list_items}
\title{Add items to a list object}
\usage{
add_list_items(obj, ..., overwrite = TRUE, remove_null = FALSE)
}
\arguments{
\item{obj}{A list object}

\item{...}{Optional named or unnamed arguments which should be appended
as new items to \code{obj}.}

\item{overwrite}{A logical value, defining if existing entries should be
overwritten (\code{overwrite = TRUE}) or not.}

\item{remove_null}{A logical value, defining if for the resulting list object
all \code{NULL} values should be stripped.}
}
\value{
The modified list object, holding additional list entries.
}
\description{
Add items to a list object
}
