% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{calc_date}
\alias{calc_date}
\title{Calculate a date vector}
\usage{
calc_date(day, month, year)
}
\arguments{
\item{day}{A numeric vector holding the day number (1 - 31) of the month}

\item{month}{A numeric vector holding the month numbers (1 - 12)}

\item{year}{A numeric vector holding the year numbers (>= 2000)}
}
\description{
Calculate a date vector
}
