% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{calc_difftime_days}
\alias{calc_difftime_days}
\alias{calc_difftime_years}
\title{Calculate time difference between two dates}
\usage{
calc_difftime_days(start, end)

calc_difftime_years(start, end)
}
\arguments{
\item{start}{A date vector holding the start dates of the time interval.}

\item{end}{A date vector holding the end dates of the time interval.
Must be of the same length as \code{start}.}
}
\description{
The following functions can be used to calculate the time difference between
two points of time:
\itemize{
\item \code{calc_difftime_days()}: Calculates the number of days between two dates
(e.g. \code{365} for \verb{1.1.2019} and \verb{1.1.2020})
\item \code{calc_difftime_years()}: Calculates the number of years (including fractional years) between
to dates. This function also takes leap years into account.
(e.g.: \code{1.000} for \verb{1.3.2019} and \verb{1.3.2020})
}
}
