% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{calc_difftime_days_helper}
\alias{calc_difftime_days_helper}
\title{Helper function for calculating the time difference in days}
\usage{
calc_difftime_days_helper(start, end, err_h)
}
\arguments{
\item{start}{A date vector holding the start dates of the time interval.}

\item{end}{A date vector holding the end dates of the time interval.
Must be of the same length as \code{start}.}

\item{err_h}{Error handling function.}
}
\description{
Helper function for calculating the time difference in days
}
