% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_atomic}
\alias{check_atomic}
\alias{check_numeric}
\title{Useful argument checks with error handling}
\usage{
check_atomic(
  obj,
  err_h = composerr(),
  allow_null = FALSE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  obj_name = NULL
)

check_numeric(
  obj,
  err_h,
  min_val = -Inf,
  max_val = Inf,
  allow_null = FALSE,
  allow_inf = TRUE,
  allow_na = FALSE,
  allow_nan = FALSE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  obj_name = NULL
)
}
\arguments{
\item{obj}{The object that should be checked}

\item{err_h}{An error handling function, which will be called, when the
check fails.}

\item{allow_null}{A logical value, defining if \code{NULL} is a allowed value for \code{obj}.}

\item{min_len}{A number, defining the minimal length of the vector \code{obj}.}

\item{max_len}{A number, defining the maximal length of the vector \code{obj}.}

\item{len}{A number, defining the exact length of the vector \code{obj}.}

\item{obj_name}{An optional string holding the name of the \code{obj}, which
should be used for referring to in the error message.}

\item{min_val}{A number or \verb{+/-Inf} used for the check: \code{obj >= min_val}.}

\item{max_val}{A number or \verb{+/-Inf} used for the check: \code{obj <= max_val}.}

\item{allow_inf}{A logical value, defining if \verb{+/-Inf} are allowed values for \code{obj}.}

\item{allow_na}{A logical value, defining if \code{NA} is allowed for \code{obj}.}

\item{allow_nan}{A logical value, defining if \code{NaN} is allowed for \code{obj}.}
}
\value{
The object given in \code{obj}
}
\description{
Useful argument checks with error handling
}
