% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_wide_convertor.R
\name{convert2long}
\alias{convert2long}
\title{convert2long}
\usage{
convert2long(expression_data, sample_info, variable_info)
}
\arguments{
\item{expression_data}{expression_data}

\item{sample_info}{sample_info}

\item{variable_info}{variable_info}
}
\description{
Convert from wide data to long data.
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
