% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_wide_convertor.R
\name{convert2wide}
\alias{convert2wide}
\title{convert2wide}
\usage{
convert2wide(
  data,
  sample_info_name = c("sample_id"),
  variable_info_name = c("variable_id")
)
}
\arguments{
\item{data}{data}

\item{sample_info_name}{sample_info_name}

\item{variable_info_name}{variable_info_name}
}
\description{
Convert from wide data to long data.
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
