% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_d_vars.R
\name{create_d_vars}
\alias{create_d_vars}
\alias{create_d_vars_all}
\title{Calculate and add difference-variables for count variables in data.frames}
\usage{
create_d_vars(x, n_var, order_var = NULL, grouping_var = NULL)

create_d_vars_all(x, order_var = NULL, grouping_var = NULL, except_vars = NULL)
}
\arguments{
\item{x}{A data.frame holding the data}

\item{n_var}{A character vector holding the names of the counting variables
for which the difference variables should be created.}

\item{order_var}{An optional string holding the name of the variable, which
is used for ordering the data.frame. If omitted, the
pre-sorting is omitted.}

\item{grouping_var}{An optional string holding the name of the grouping variable.
This variable is used for grouping the data.frame. If omitted, the
pre-grouping is omitted.}

\item{except_vars}{A character vector holding variables, which should
not be used for longitudinal calculation.}
}
\value{
A modified data.frame, holding the calculated variables.
}
\description{
Often we have data.frames which have counting variables like
\itemize{
\item \code{n_geb}: number of born children
\item \code{n_gest}: number of died people
\item \code{n_pop}: number of people living in Tyrol
\item etc.
}
}
\details{
For this variables we often want to calculate the variables:
\itemize{
\item \verb{d_*} (e.g. \code{d_geb}): The absolute difference (growth) compared to the previous year.
\item \verb{da_*} (e.g. \code{da_geb}): The relative difference (relative growth) compared to the previous year.
}

For this purpose we can have:
\itemize{
\item an ordering prior to the calcuation: often \code{order = jahr}
\item an optional grouping prior to the calculation (e.g. by \code{geschl} or some other grouping var)
}

The following functions can do this calculations:
\itemize{
\item \code{create_d_vars()}: Calculate \verb{d_*} and \verb{da_*} for a set of given variables (mostly of type \verb{n_*})
\item \code{create_d_vars_all()}: Calculate \verb{d_*} and \verb{da_*} for all variables except of a few given variables.
}
}
