% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_d_vars.R
\name{create_d_vars_helper}
\alias{create_d_vars_helper}
\title{helper function for \code{\link[=create_d_vars]{create_d_vars()}}}
\usage{
create_d_vars_helper(
  x,
  n_var,
  order_var = NULL,
  grouping_var = NULL,
  err_h = composerr("Error while calling 'create_d_vars_helper()': ")
)
}
\arguments{
\item{x}{A data.frame holding the data}

\item{n_var}{A character vector holding the names of the counting variables
for which the difference variables should be created.}

\item{order_var}{An optional string holding the name of the variable, which
is used for ordering the data.frame. If omitted, the
pre-sorting is omitted.}

\item{grouping_var}{An optional string holding the name of the grouping variable.
This variable is used for grouping the data.frame. If omitted, the
pre-grouping is omitted.}

\item{err_h}{Error handling function.}
}
\description{
helper function for \code{\link[=create_d_vars]{create_d_vars()}}
}
