% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{date_2_year}
\alias{date_2_year}
\title{Extract year number from date}
\usage{
date_2_year(date)
}
\arguments{
\item{date}{A date vector}
}
\value{
The year number vector
}
\description{
Extract year number from date
}
