% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_manipulation.R
\name{eval_closure}
\alias{eval_closure}
\title{Eval code in closure without scoping problems}
\usage{
eval_closure(expr, vars = NULL, lookup_env = parent.frame())
}
\arguments{
\item{expr}{The expression, which should be evaluated inside of the
closure.}

\item{vars}{An optional character vector holding the names of the variables
which should be copied to the new function scope.}

\item{lookup_env}{The environment holding the variables given in \code{vars}.
The default is the calling environment.}
}
\description{
Eval code in closure without scoping problems
}
