% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_manipulation.R
\name{format_abs}
\alias{format_abs}
\alias{format_rel}
\title{Transform numeric vector to vector of strings}
\usage{
format_abs(x, signed = FALSE, latex_minus = FALSE, ...)

format_rel(
  x,
  signed = FALSE,
  latex_minus = FALSE,
  nsmall = 1,
  nsmall_max = 2,
  ...
)
}
\arguments{
\item{x}{A numerical vector.}

\item{signed}{A logical flag, defining if positive numbers should also have
a leading signature sign.}

\item{latex_minus}{A logical value. If set to \code{TRUE}, then possible minus
signs are replace by a special LaTeX command. If the resulting texts
are compiled by a LaTeX compile, then the resulting text contains
minus signs where no line break is possible between the minus sign and
the following number. For this, the preamble of the surrounding LaTeX
document must contain the following line of code:
\verb{\\\\usepackage[shortcuts]\{extdash\}}}

\item{...}{Various other arguments passed into the \code{\link[=format]{format()}} function.
These arguments can be used in order to change the behavior
of \code{format_abs()}, \code{format_rel()}, etc.}

\item{nsmall}{A positive integer, defining the minimum number of displayed
decimal places}

\item{nsmall_max}{A positive integer, defining the maximum number of displayed
decimal places}
}
\description{
The following functions transform numbers to strings:
\itemize{
\item \code{format_abs()}: formatted as absolute values (without decimal places)
\item \code{format_rel()}: formatted as relative values (with decimal places)
}
}
\details{
This functions can be used to create LaTeX text or normal text.
For LaTeX text the minus symbol has some extra LaTeX formatting commands.
}
