% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{get_defaults}
\alias{get_defaults}
\title{Get all arguments (including their default values) of a function}
\usage{
get_defaults(fn)
}
\arguments{
\item{fn}{A function.}
}
\value{
A list holding all fn arguments and their default values.
}
\description{
This function only works for non-primitive functions.
}
