% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4metTools.R
\name{GetMatchResult}
\alias{GetMatchResult}
\title{GetMatchResult}
\usage{
GetMatchResult(spec.exp, spec.lib, weight.int = 1, weight.mz = 0,
  ppm.ms2match = 30, mz.ppm.thr = 400, ppm.sanity.check = 100,
  is.sanity.check = FALSE, direction = direction, ...)
}
\description{
GetMatchResult
}
