% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4metTools.R
\name{GetSpec2Match}
\alias{GetSpec2Match}
\title{GetSpec2Match}
\usage{
GetSpec2Match(spec.exp, spec.lib, ppm.ms2match = 30, mz.ppm.thr = 400,
  direction = c("reverse", "forward"))
}
\description{
GetSpec2Match
}
