% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_manipulation.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Leap year checker}
\usage{
is_leap_year(x)
}
\arguments{
\item{x}{A numeric vector holding the year numbers}
}
\value{
A logical vector, where each entry displays if the year in the
corresponding \code{x} entry was a leap year.
}
\description{
Leap year checker
}
