% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_tools.R
\name{parse_copied_range}
\alias{parse_copied_range}
\title{Parse a copied range (possibly Excel) to a data.frame}
\usage{
parse_copied_range(
  x,
  col_names = NULL,
  num_cols = NULL,
  big_mark = ".",
  dec_mark = ",",
  col_sep = "\\t",
  row_sep = "\\n"
)
}
\arguments{
\item{x}{The copied string holding the values from the range}

\item{col_names}{Optional character vector holding the names which should be assigned to
the data.frame columns. If omitted, then the column names \code{X1}, \code{X2} ,... are used.}

\item{num_cols}{Optional vector holding the column numbers which should be
converted into a numeric columns. Of omitted, then all columns are assumed
to be numeric.}

\item{big_mark}{A string holding the character symbol for the big mark in the
numeric columns.}

\item{dec_mark}{A string holding the character symbol for the decimal mark.}

\item{col_sep}{A string holding the column separator.}

\item{row_sep}{A string holding the row separator.}
}
\value{
A data.frame holding the parsed data.
}
\description{
This function allows the user to copy a range (for example in Excel) and
paste it to a string variable
in R and then parse the string into a data.frame
}
