% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{set_fn_defaults}
\alias{set_fn_defaults}
\title{Overwrite the default values of a given function}
\usage{
set_fn_defaults(fn, ...)
}
\arguments{
\item{fn}{The function for which the defaults should be set.}

\item{...}{Various named functions arguments, which should be set as default}
}
\value{
A new function, with overwritten defaults
}
\description{
This functions takes a function and modifies its default values.
}
