% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_manipulation.R
\name{shift_vec}
\alias{shift_vec}
\title{Shift vector}
\usage{
shift_vec(x, by = 1L, rotate = FALSE)
}
\arguments{
\item{x}{A vector}

\item{by}{Number of positions to shift. Negative values produce a shift to
the left}

\item{rotate}{Should the lost values be inserted as input values?
If \code{rotate = FALSE}, then \code{NA} will be inserted instead.}
}
\description{
Shift vector
}
