% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_manipulation.R
\name{stringify}
\alias{stringify}
\title{Stringify vector}
\usage{
stringify(
  x,
  str_collapse = ", ",
  quote = TRUE,
  str_before = "",
  str_after = "",
  new_line = FALSE
)
}
\arguments{
\item{x}{A vector}

\item{str_collapse}{A string, used for separating each vector entry.}

\item{quote}{A logical. If set to \code{TRUE}, then each entry of \code{x} is quoted with \verb{'}.}

\item{str_before}{A string placed before each vector entry.}

\item{str_after}{A string placed after each vector entry.}

\item{new_line}{A logical. If set to \code{TRUE}, then a new line command is
placed at the end of the resulting string.}
}
\value{
A string showing the entries of the vector.
}
\description{
Stringify vector
}
