% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytools_logo.R
\name{tinytools_logo}
\alias{tinytools_logo}
\title{tinytools_logo}
\usage{
tinytools_logo()
}
\description{
Get the detailed of metPath package.
}
\examples{
tinytools_logo()
}
\author{
Xiaotao Shen
\email{shenxt@stanford.edu}
}
