% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{tinyTools_packages}
\alias{tinyTools_packages}
\title{List all packages in the tinyTools}
\usage{
tinyTools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tinyTools in the list?}
}
\description{
List all packages in the tinyTools
}
\examples{
tinyTools_packages()
}
