% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_tools.R
\name{ttable}
\alias{ttable}
\title{The function \code{table} with \code{useNA = "always"}}
\usage{
ttable(...)
}
\arguments{
\item{...}{All arguments for \code{table}}
}
\description{
The function \code{table} with \code{useNA = "always"}
}
