% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4metTools.R
\name{TuneMS2}
\alias{TuneMS2}
\title{TuneMS2}
\usage{
TuneMS2(spec, mz.precursor, is.include.precursor = TRUE, snthr = 3,
  noise.ms2 = 3, mz.range.ms2 = NULL, int.ms2.min.abs,
  int.ms2.min.relative = 0.01, is.apply.ms2.min.relative = TRUE,
  is.check.sanity = TRUE, int.check.sanity = 50,
  ppm.precursor.filter = 20, ...)
}
\description{
TuneMS2
}
