% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_numeric}
\alias{validate_numeric}
\alias{validate_integer}
\alias{validate_logical}
\alias{validate_character}
\alias{validate_vector}
\alias{validate_list}
\title{Useful validations with proper error handling}
\usage{
validate_numeric(
  obj,
  err_h = composerr(),
  min_val = -Inf,
  max_val = Inf,
  allow_null = FALSE,
  allow_inf = TRUE,
  allow_na = FALSE,
  allow_nan = FALSE,
  allow_duplicates = TRUE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  named = NULL,
  element_names = NULL,
  allowed_names = NULL,
  required_names = NULL,
  forbidden_names = NULL,
  obj_name = NULL
)

validate_integer(
  obj,
  err_h = composerr(),
  min_val = -Inf,
  max_val = Inf,
  allowed_values = NULL,
  allow_null = FALSE,
  allow_inf = TRUE,
  allow_na = FALSE,
  allow_nan = FALSE,
  allow_duplicates = TRUE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  named = NULL,
  element_names = NULL,
  allowed_names = NULL,
  required_names = NULL,
  forbidden_names = NULL,
  obj_name = NULL
)

validate_logical(
  obj,
  err_h = composerr(),
  allow_null = FALSE,
  allow_na = FALSE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  named = NULL,
  element_names = NULL,
  allowed_names = NULL,
  required_names = NULL,
  forbidden_names = NULL,
  obj_name = NULL
)

validate_character(
  obj,
  err_h = composerr(),
  allowed_values = NULL,
  allow_null = FALSE,
  allow_na = FALSE,
  allow_duplicates = TRUE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  named = NULL,
  element_names = NULL,
  allowed_names = NULL,
  required_names = NULL,
  forbidden_names = NULL,
  obj_name = NULL
)

validate_vector(
  obj,
  err_h = composerr(),
  allow_null = FALSE,
  allow_na = FALSE,
  allow_duplicates = TRUE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  named = NULL,
  element_names = NULL,
  allowed_names = NULL,
  required_names = NULL,
  forbidden_names = NULL,
  obj_name = NULL
)

validate_list(
  obj,
  err_h,
  allow_null = FALSE,
  min_len = 0L,
  max_len = Inf,
  len = NULL,
  named = NULL,
  element_names = NULL,
  allowed_names = NULL,
  required_names = NULL,
  forbidden_names = NULL,
  obj_name = NULL
)
}
\arguments{
\item{obj}{The object that should be checked}

\item{err_h}{An error handling function created with
\link[composerr:composerr]{composerr()}. This handler will be called, if the
validation fails.}

\item{min_val}{A number or \verb{+/-Inf} used for the check: \code{obj >= min_val}.}

\item{max_val}{A number or \verb{+/-Inf} used for the check: \code{obj <= max_val}.}

\item{allow_null}{A logical value, defining if \code{NULL} is a allowed value for \code{obj}.}

\item{allow_inf}{A logical value, defining if \verb{+/-Inf} are allowed values for \code{obj}.}

\item{allow_na}{A logical value, defining if \code{NA} is allowed for \code{obj}.}

\item{allow_nan}{A logical value, defining if \code{NaN} is allowed for \code{obj}.}

\item{allow_duplicates}{A logical value, defining if duplicated
vector elements are allowed for \code{obj}.}

\item{min_len}{A number, defining the minimal length of the vector \code{obj}.}

\item{max_len}{A number, defining the maximal length of the vector \code{obj}.}

\item{len}{A number, defining the exact length of the vector \code{obj}.}

\item{named}{A logical value, defining if \code{obj} is required to be a named
vector (\code{named = TRUE}) or if \code{obj} is required to be an unnamed vector
(\code{named = FALSE}). If it does not matter wether \code{obj} is named or unnamed
then one should set \code{named = NULL}.}

\item{element_names}{Optional character vector holding the required names
for the vector elements. The order of the named elements does not matter.}

\item{allowed_names}{Optional character vector holding the allowed names
for the vector elements.}

\item{required_names}{Optional character vector holding the required names
for the vector elements.}

\item{forbidden_names}{Optional character vector holding the forbidden names
for the vector elements.}

\item{obj_name}{An optional string holding the name of the \code{obj}, which
should be used for referring to in the error message.}

\item{allowed_values}{An optional vector holding allowed values for the
vector elements.}
}
\value{
The object given in \code{obj}
}
\description{
Useful validations with proper error handling
}
