% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4metTools.R
\name{getSpectraMatchScore}
\alias{getSpectraMatchScore}
\title{getSpectraMatchScore}
\usage{
getSpectraMatchScore(
  exp.spectrum,
  lib.spectrum,
  ppm.tol = 30,
  mz.ppm.thr = 400,
  fraction.weight = 0.2,
  dp.forward.weight = 0.7,
  dp.reverse.weight = 0.1
)
}
\description{
getSpectraMatchScore
}
