% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{set_fn_defaults}
\alias{set_fn_defaults}
\title{Overwrite the default values of a given function}
\usage{
set_fn_defaults(fn, ...)
}
\arguments{
\item{fn}{The function for which the defaults should be set.}

\item{...}{Various named functions arguments, which should be set as default}
}
\value{
A new function, with overwritten defaults
}
\description{
This functions takes a function and modifies its default values.
TODO: FIX \strong{ERROR}, that calling \code{fn_new(1)} does not work,
but \code{fn_new(x = 1)} does. This means, that piping does not
work any more, since it uses unnamed caller args.
}
