% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tinytools_packages}
\alias{tinytools_packages}
\title{List all packages in the tinytools}
\usage{
tinytools_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tinytools in the list?}
}
\description{
List all packages in the tinytools
}
\examples{
tinytools_packages()
}
